+++ opencog/atoms/foreign/DatalogAST.cc
@@ -57,138 +57,6 @@ DatalogAST::DatalogAST(const std::string& sexpr)
 	parse(sexpr);
 }
 
-void DatalogAST::parse(const std::string& sexpr)
-{
-	size_t l = 0;
-	size_t r = 0;
-	while (std::string::npos != r)
-	{
-		Handle h(get_next_expr(sexpr, l, r));
-		_outgoing.emplace_back(h);
-printf("duuude made %s\n", h->to_short_string().c_str());
-	}
-}
-
-// ---------------------------------------------------------------
-
-// Parse expressions such as
-// likes(John, Mary).
-Handle DatalogAST::get_next_expr(const std::string& sexpr, size_t& l, size_t &r)
-{
-	l = sexpr.find_first_not_of(" \t\n", l);
-	if (std::string::npos == l)
-		throw SyntaxException(TRACE_INFO, "Unexpected blank line");
-
-	r = sexpr.find_first_of("( \t\n", l+1);
-	const std::string& spred = sexpr.substr(l, r-l);
-printf("duuude got pred >>%s<<\n", spred.c_str());
-	Handle pred = createNode(PREDICATE_NODE, std::move(spred));
-
-	l = sexpr.find_first_not_of(" \t\n", r);
-	if (std::string::npos == l or '(' != sexpr[l])
-		throw SyntaxException(TRACE_INFO, "Expecting open-paren");
-
-	l++; // step past open paren
-
-	HandleSeq clist;
-	while (std::string::npos != l and '.' != sexpr[l])
-	{
-		r = sexpr.find_first_of(",) \t\n", l);
-		const std::string& cept = sexpr.substr(l, r-l);
-printf("duuude got concept >>%s<<\n", cept.c_str());
-
-		clist.emplace_back(createNode(CONCEPT_NODE, cept));
-		if (')' == sexpr[r]) break;
-		l = sexpr.find_first_not_of(", \t\n", r);
-	}
-
-	l = sexpr.find_first_of(".", l);
-	if (std::string::npos == l)
-		throw SyntaxException(TRACE_INFO, "Expecting terminating period");
-
-	l++;
-	r = sexpr.find_first_not_of(" \t\n", l);
-
-	Handle evl = createLink(EVALUATION_LINK,
-		pred, createLink(std::move(clist), LIST_LINK));
-	return evl;
-}
-
-// ---------------------------------------------------------------
-
-std::string DatalogAST::prt_datalog(const Handle& h)
-{
-	Type t = h->get_type();
-	if (h->is_node())
-		return h->get_name();
-
-	std::string rv = "";
-	if (DATALOG_AST == t)
-	{
-		rv += "(DatalogAst \"";
-		for (const Handle& ho: h->getOutgoingSet())
-			rv += prt_datalog(ho);
-		rv += "\")";
-	}
-	else if (EVALUATION_LINK == t)
-	{
-		// First comes a PredicateNode
-		rv += prt_datalog(h->getOutgoingAtom(0));
-
-		rv += "(";
-		// Next comes a ListLink
-		const Handle& ll = h->getOutgoingAtom(1);
-		for (const Handle& ho : ll->getOutgoingSet())
-		{
-			rv += prt_datalog(ho);
-			rv += ", ";
-		}
-		rv.pop_back();
-		rv.pop_back();
-		rv += ").";
-	}
-	else
-		rv += "foo";
-
-	return rv;
-}
-
-// ---------------------------------------------------------------
-
-std::string DatalogAST::to_string(const std::string& indent) const
-{
-	if (0 == _outgoing.size())
-		return indent + "(DatalogAst \"" + _name + "\") ; " + id_to_string();
-
-	std::string rv = indent + "(DatalogAst\n";
-	for (const Handle& h: _outgoing)
-		rv += h->to_string(indent + "  ") + "\n";
-
-	rv += indent + ") ; " + id_to_string();
-	return rv;
-}
-
-std::string DatalogAST::to_short_string(const std::string& indent) const
-{
-	if (0 == indent.size())
-		return _name + "\n" + to_short_string(";") + "\n";
-
-	// Debugging print
-	if (0 == _outgoing.size()) // this should never happen
-		return _name + "XXX-borken";
-
-	std::string rv = "";
-	for (const Handle& h: _outgoing)
-	{
-		if (DATALOG_AST == h->get_type())
-			rv += h->to_short_string("xx") + " ";
-		else
-			rv += indent + h->to_short_string(indent);
-	}
-
-	return rv;
-}
-
 // ---------------------------------------------------------------
 // Custom factory, because its a hermaphrodite. The ForgeinAST will
 // pass us a string, behaving like a node, which we parse into an
